ControlPanel/requirements.txt
pip==25.2
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
pyserial==3.5